% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.ab_h}
\alias{.ab_h}
\title{Internal function to generate constrained a's and b's}
\usage{
.ab_h(n_ts)
}
\arguments{
\item{n_ts}{Number of time series}
}
\description{
Function to generate individual a_h and b_h of lenght n_ts that comply with a_h + b_h < 1
}
\author{
Philippe Rast
}
