% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.covranef}
\alias{.covranef}
\title{Create random effects in fixed correlation}
\usage{
.covranef(fixed_covmat, ranef_sd)
}
\arguments{
\item{fixed_covmat}{Fixed covariance matrix, to which ranefs will be added}

\item{ranef_sd}{SD of random effect that will be added to the cholesky transformed values}
}
\description{
Simulate times series
}
\author{
Philippe Rast
}
