% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simuDCC}
\alias{.simuDCC}
\title{Simulate TS's for N individuals}
\usage{
.simuDCC(
  tslength,
  n_ts,
  N,
  phi0_fixed = rep(0, n_ts),
  phi0_sd = 0.5,
  log_c_fixed = rep(0, n_ts),
  log_c_r_sd = 0.1,
  a_h_fixed = rep(0, n_ts),
  a_h_r_sd = 0.1,
  b_h_fixed = rep(0, n_ts),
  b_h_r_sd = 0.1,
  l_a_q_fixed = 0,
  l_a_q_r_sd = 0.1,
  l_b_q_fixed = 0,
  l_b_q_r_sd = 0.1,
  ranef_sd_S,
  alpha = 0.5
)
}
\arguments{
\item{tslength}{Length of time series, for all N (everyone has same TS length)}

\item{n_ts}{Number of simultaneous TS per person. Currently only supports 3 or less}

\item{N}{Subjects}

\item{phi0_fixed}{Vector of population values of length n_ts}

\item{ranef_sd_S}{Size of random effects SD in the function that generates random effects around fixed sqrt(r). I'd suggest a rather small number}
}
\description{
Simulate TS's for N individuals
}
