% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcnet.R
\name{standat}
\alias{standat}
\title{Standardize input data to facilitate computation}
\usage{
standat(
  data,
  J,
  group,
  xC,
  P,
  Q,
  standardize_data,
  distribution,
  meanstructure
)
}
\arguments{
\item{data}{Time-series data}

\item{J}{numeric. Number of groups (individuals)}

\item{group}{Vector with group id of full length}

\item{xC}{Numeric vector or matrix.}

\item{P}{Numeric.}

\item{Q}{Numeric.}

\item{standardize_data}{Logical.}

\item{distribution}{Character.}

\item{meanstructure}{Character.}
}
\value{
dcnet stan data list.
}
\description{
Standardize input data to facilitate computation
}
\keyword{internal}
