% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcnet.R
\docType{data}
\name{supported_models}
\alias{supported_models}
\title{Models supported by dcnet}
\format{
An object of class \code{character} of length 1.
}
\usage{
supported_models
}
\description{
To be used when checking whether a parameterization or object type is a supported type.
May facilitate more parameterizations, as we only have to update these, and the switch statements.
}
\author{
Philippe Rast
}
\keyword{internal}
