% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simuDCC}
\alias{.simuDCC}
\title{Simulate TS's for N individuals}
\usage{
.simuDCC(tslength, n_ts, N, ranef_sd_S)
}
\arguments{
\item{tslength}{Length of time series, for all N (everyone has same TS length)}

\item{n_ts}{Number of simultaneous TS per person. Currently only supports 3 or less}

\item{N}{Subjects}

\item{ranef_sd_S}{Size of random effects SD in the function that generates random effects around fixed sqrt(r). I'd suggest a rather small number}
}
\description{
Simulate TS's for N individuals
}
